requirements.txt
Flask
flask_SqlAlchemy
PyMysql 
Pandas
NumPy
matplotlib
